feat: separate out widgets instead of widget return methods (this is an anti pattern)
